Refactor the Base.js file
In order to ease maintainability and future tests, we split one big file
into multiple tiny ones.
