+++ src/abstracts/Base/options.js
@@ -1,0 +1,49 @@
+import merge from 'deepmerge';
+
+/**
+ * Get a component's options.
+ *
+ * @param  {Base}        instance The component's instance.
+ * @param  {HTMLElement} element  The component's root element.
+ * @param  {Object}      config   The component's default config.
+ * @return {Object}               The component's merged options.
+ */
+export function getOptions(instance, element, config) {
+  let options = {};
+  if (element.dataset.options) {
+    try {
+      options = JSON.parse(element.dataset.options);
+    } catch (err) {
+      throw new Error('Can not parse the `data-options` attribute. Is it a valid JSON string?');
+    }
+  }
+
+  options = merge(config, options);
+  instance.$emit('get:options', options);
+  return options;
+}
+
+/**
+ * Set a component instance options.
+ *
+ * @param {Base}        instance   The component's instance.
+ * @param {HTMLElement} element    The component's root element.
+ * @param {Object}      newOptions The new options object.
+ */
+export function setOptions(instance, element, newOptions) {
+  let options = {};
+  if (element.dataset.options) {
+    try {
+      options = JSON.parse(element.dataset.options);
+    } catch (err) {
+      throw new Error('Can not parse the `data-options` attribute. Is it a valid JSON string?');
+    }
+  }
+  options = merge(options, newOptions);
+  element.dataset.options = JSON.stringify(options);
+}
+
+export default {
+  getOptions,
+  setOptions,
+};
